refactor: cardProps를 관리하는 파일 추가
