+++ src/types/card.ts
@@ -1,0 +1,9 @@
+interface CardProps {
+  id: string;
+  thumbnail: string;
+  title: string;
+  description: string;
+  link: string;
+}
+
+export default CardProps;
